[NFC] Split VectorExt attributes into a separate file
In preparation for potentially adding more attribute definitions.
