+++ llvm-external-projects/iree-dialects/BUILD.bazel
@@ -648,22 +648,6 @@ gentbl_cc_library(
             "include/iree-dialects/Dialect/VectorExt/IR/VectorExtDialect.cpp.inc",
         ),
         (
-            ["--gen-attrdef-decls"],
-            "include/iree-dialects/Dialect/VectorExt/IR/VectorExtAttrs.h.inc",
-        ),
-        (
-            ["--gen-attrdef-defs"],
-            "include/iree-dialects/Dialect/VectorExt/IR/VectorExtAttrs.cpp.inc",
-        ),
-        (
-            ["--gen-enum-decls"],
-            "include/iree-dialects/Dialect/VectorExt/IR/VectorExtEnums.h.inc",
-        ),
-        (
-            ["--gen-enum-defs"],
-            "include/iree-dialects/Dialect/VectorExt/IR/VectorExtEnums.cpp.inc",
-        ),
-        (
             ["--gen-op-decls"],
             "include/iree-dialects/Dialect/VectorExt/IR/VectorExtOps.h.inc",
         ),
@@ -682,6 +666,34 @@ gentbl_cc_library(
     ],
     tblgen = "@llvm-project//mlir:mlir-tblgen",
     td_file = "include/iree-dialects/Dialect/VectorExt/IR/VectorExtOps.td",
+    deps = [
+        ":TdFiles",
+    ],
+)
+
+gentbl_cc_library(
+    name = "IREEVectorExtAttrsIncGen",
+    strip_include_prefix = "include",
+    tbl_outs = [
+        (
+            ["--gen-attrdef-decls"],
+            "include/iree-dialects/Dialect/VectorExt/IR/VectorExtAttrs.h.inc",
+        ),
+        (
+            ["--gen-attrdef-defs"],
+            "include/iree-dialects/Dialect/VectorExt/IR/VectorExtAttrs.cpp.inc",
+        ),
+        (
+            ["--gen-enum-decls"],
+            "include/iree-dialects/Dialect/VectorExt/IR/VectorExtEnums.h.inc",
+        ),
+        (
+            ["--gen-enum-defs"],
+            "include/iree-dialects/Dialect/VectorExt/IR/VectorExtEnums.cpp.inc",
+        ),
+    ],
+    tblgen = "@llvm-project//mlir:mlir-tblgen",
+    td_file = "include/iree-dialects/Dialect/VectorExt/IR/VectorExtAttrs.td",
     deps = [
         ":TdFiles",
     ],
@@ -717,6 +729,7 @@ cc_library(
     ]),
     includes = ["include"],
     deps = [
+        ":IREEVectorExtAttrsIncGen",
         ":IREEVectorExtIncGen",
         ":IREEVectorExtInterfacesIncGen",
         "@llvm-project//llvm:Support",
