+++ llvm-external-projects/iree-dialects/include/iree-dialects/Dialect/VectorExt/IR/CMakeLists.txt
@@ -8,10 +8,6 @@ endfunction()
 
 function(_add_dialect)
   set(LLVM_TARGET_DEFINITIONS VectorExtOps.td)
-  mlir_tablegen(VectorExtAttrs.h.inc -gen-attrdef-decls)
-  mlir_tablegen(VectorExtAttrs.cpp.inc -gen-attrdef-defs)
-  mlir_tablegen(VectorExtEnums.h.inc -gen-enum-decls)
-  mlir_tablegen(VectorExtEnums.cpp.inc -gen-enum-defs)
   mlir_tablegen(VectorExtOps.h.inc -gen-op-decls)
   mlir_tablegen(VectorExtOps.cpp.inc -gen-op-defs)
   mlir_tablegen(VectorExtTypes.h.inc -gen-typedef-decls)
@@ -20,6 +16,14 @@ function(_add_dialect)
   mlir_tablegen(VectorExtDialect.cpp.inc --gen-dialect-defs --dialect=iree_vector_ext)
   add_public_tablegen_target(IREEVectorExtIncGen)
   add_dependencies(mlir-headers IREEVectorExtIncGen)
+
+  set(LLVM_TARGET_DEFINITIONS VectorExtAttrs.td)
+  mlir_tablegen(VectorExtAttrs.h.inc -gen-attrdef-decls)
+  mlir_tablegen(VectorExtAttrs.cpp.inc -gen-attrdef-defs)
+  mlir_tablegen(VectorExtEnums.h.inc -gen-enum-decls)
+  mlir_tablegen(VectorExtEnums.cpp.inc -gen-enum-defs)
+  add_public_tablegen_target(IREEVectorExtAttrsIncGen)
+  add_dependencies(mlir-headers IREEVectorExtAttrsIncGen)
 endfunction()
 
 function(_add_doc)
