+++ llvm-external-projects/iree-dialects/lib/Dialect/VectorExt/IR/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_mlir_library(IREEVectorExtDialect
+  VectorExtAttrs.cpp
   VectorExtDialect.cpp
   VectorExtOps.cpp
 
@@ -6,6 +7,7 @@ add_mlir_library(IREEVectorExtDialect
   ${IREE_DIALECTS_SOURCE_DIR}/include
 
   DEPENDS
+  IREEVectorExtAttrsIncGen
   IREEVectorExtIncGen
 
   LINK_LIBS PUBLIC
